backend/requirements.txt
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
fastapi==0.110.0
python-dotenv==1.0.1
requests==2.31.0
uvicorn==0.27.1